integration test for api filter endpoints

+++ source/informationHandler.cpp
@@ -66,6 +66,10 @@ int InformationHandler::edit(const json& info, const std::string& id) {
 }
 
 int InformationHandler::remove(const std::string& id) {
+    //Check if id has length of an uuid
+    if (id.size() != 36) {
+        return 400; // Bad Request
+    }
     try {
         const std::string informationFilePath = localEyeCANPath + id + ".json";
         if (std::filesystem::remove(informationFilePath)) {
